+++ test/unit/fuzz-tools.test.js
@@ -340,6 +340,7 @@ import * as kwbRetinopathyV360 from '../../lib/kwb-retinopathy-v360.js';
 import * as tannerStagingV361 from '../../lib/tanner-staging-v361.js';
 import * as forresterHemodynamicV362 from '../../lib/forrester-hemodynamic-v362.js';
 import * as shafferAngleV363 from '../../lib/shaffer-angle-v363.js';
+import * as casTedV364 from '../../lib/cas-ted-v364.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -625,6 +626,7 @@ const MODULES = {
   'tanner-staging-v361.js': tannerStagingV361,
   'forrester-hemodynamic-v362.js': forresterHemodynamicV362,
   'shaffer-angle-v363.js': shafferAngleV363,
+  'cas-ted-v364.js': casTedV364,
   'tb-testing.js': tbTesting,
 };